+++ mangadex_library/lib/mangadex_library.dart
@@ -437,7 +437,7 @@ Future<MangaReadingStatus> getMangaReadingStatus(
 
 Future<ResultOk> setMangaReadingStatus(
     String token, String mangaId, ReadingStatus? status) async {
-  var statusString = status != null ? parseStatusFromEnum(status) : '';
+  var statusString = status != null ? parseStatusFromEnum(status) : 'reading';
   var unencodedPath = '/manga/$mangaId/status';
   final uri = 'https://$authority$unencodedPath';
   var response = await http.post(Uri.parse(uri),
@@ -489,11 +489,6 @@ Future<MangaCheck> followManga(String token, String mangaId,
 Future<MangaCheck> unfollowManga(String token, String mangaId) async {
   var unencodedPath = '/manga/$mangaId/unfollow';
   final uri = 'https://$authority$unencodedPath';
-  try {
-    await removeMangaReadingStatus(token, mangaId);
-  } catch (e) {
-    print("couldn't remove manga reading status for manga $mangaId");
-  }
   var response = await http.delete(Uri.parse(uri), headers: {
     HttpHeaders.contentTypeHeader: 'application/json',
     HttpHeaders.authorizationHeader: 'Bearer $token'